requirements.txt
requests~=2.24.0
bs4~=0.0.1
beautifulsoup4~=4.9.2
lxml
praw
BotUtils
tabulate
PyGithub
discord.py